+++ Exercism/c/allergies/src/allergies.h
@@ -1,0 +1,27 @@
+#ifndef ALLERGIES_H
+#define ALLERGIES_H
+
+#include <stdbool.h>
+
+typedef enum {
+   Allergen_Eggs = 0,
+   Allergen_Peanuts,
+   Allergen_Shellfish,
+   Allergen_Strawberries,
+   Allergen_Tomatoes,
+   Allergen_Chocolate,
+   Allergen_Pollen,
+   Allergen_Cats,
+   Allergen_Count
+} Allergen_t;
+
+typedef struct {
+   int count;
+   Allergen_t *allergens;
+} Allergen_List_t;
+
+bool is_allergic_to(Allergen_t gen, int score);
+
+void get_allergens(int score, Allergen_List_t* list);
+
+#endif